\name{internals}
\alias{setGenericVerif}

\title{Set a generic inter-class function}

\description{
  Internal use only. Do not bother with this method.

  \code{setGenericVerif} registers a generic class function.
}

\usage{
  setGenericVerif(x, y)
}

\arguments{
  \item{x}{name of the function to register}

  \item{y}{signature of the function to register}
}
